pyproject.toml
[tool.poetry]
name = "foxesstools"
version = "0.1.0"
description = ""
authors = ["Jason Webb <jwebb@tuffmail.co.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.8.0"
isort = "^5.10.1"
pylint = "^2.15.2"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
